nicer error message for failed backfill
Many checkpoint sources don't support block download

+++ beacon_chain/spec/eth2_apis/eth2_rest_serialization.nim
@@ -73,7 +73,7 @@ type
   EmptyBody* = object
 
   RestGenericError* = object
-    code*: uint64
+    code*: int
     message*: string
     stacktraces*: Option[seq[string]]
